refactoring protocol manager
